Move tenant-selection bootstrap from OrgMenu into TenantGuard
Tenant selection is global app state, but it was being initialized inside OrgMenu, a presentational nav component. That made correctness depend on the menu's mount lifecycle. Extract the effect into useInitializeSelectedTenant and call it from TenantGuard, which only renders the app for users with tenant access. OrgMenu becomes a pure consumer of selectedTenant. Behavior is unchanged.
